add: Update README with additional shields and metrics
This change adds two new shields to the project's README:
Code style (Black) and SLOC, including both COCOMO and Code Badges.
It also includes a visitors badge to show the number of unique
visitors to the repository. The changes are intended to improve the
overall appearance and informational value of the README file.